Split out SolrFetcher & PySolrFetcher
