+++ harvester/fetcher/__init__.py
@@ -17,8 +17,6 @@ import urllib
 import requests
 import logbook
 from logbook import FileHandler
-import solr
-import pysolr
 from pymarc import MARCReader
 import pymarc
 import pynux.utils
@@ -33,6 +31,8 @@ from deepharvest.deepharvest_nuxeo import DeepHarvestNuxeo
 from .fetcher import Fetcher
 from .fetcher import NoRecordsFetchedException
 from .oaifetcher import OAIFetcher
+from .solrfetcher import SolrFetcher
+from .solrfetcher import PySolrFetcher
 
 EMAIL_RETURN_ADDRESS = os.environ.get('EMAIL_RETURN_ADDRESS',
                                       'example@example.com')
@@ -59,67 +59,6 @@ NUXEO_S3_THUMB_URL_FORMAT = "https://s3.amazonaws.com/static.ucldc.cdlib.org/ucl
 ###        Set can be only one json object.
 ###        '''
 ###        raise NotImplementedError
-
-
-class SolrFetcher(Fetcher):
-    def __init__(self, url_harvest, query, **query_params):
-        super(SolrFetcher, self).__init__(url_harvest, query)
-        self.solr = solr.Solr(url_harvest)  # , debug=True)
-        self.query = query
-        self.resp = self.solr.select(self.query)
-        self.numFound = self.resp.numFound
-        self.index = 0
-
-    def next(self):
-        if self.index < len(self.resp.results):
-            self.index += 1
-            return self.resp.results[self.index-1]
-        self.index = 1
-        self.resp = self.resp.next_batch()
-        if not len(self.resp.results):
-            raise StopIteration
-        return self.resp.results[self.index-1]
-
-
-class PySolrFetcher(Fetcher):
-    def __init__(self, url_harvest, query, **query_params):
-        super(PySolrFetcher, self).__init__(url_harvest, query)
-        self.solr = pysolr.Solr(url_harvest, timeout=1)
-        self.queryParams = {'q':query, 'sort':'id asc', 'wt':'json',
-                'cursorMark':'*'}
-        self.get_next_results()
-        self.numFound = self.results['response'].get('numFound')
-        self.index = 0
-
-    def set_select_path(self):
-        '''Set the encoded path to send to Solr'''
-        queryParams_encoded = pysolr.safe_urlencode(self.queryParams)
-        self.selectPath = 'query?{}'.format(queryParams_encoded)
-
-    def get_next_results(self):
-        self.set_select_path()
-        resp = self.solr._send_request('get', path=self.selectPath)
-        self.results = self.solr.decoder.decode(resp)
-        self.nextCursorMark = self.results.get('nextCursorMark')
-        self.iter = self.results['response']['docs'].__iter__()
-
-    def next(self):
-        try:
-            next_result = self.iter.next()
-            self.index += 1
-            return next_result
-        except StopIteration:
-            if self.index >= self.numFound:
-                raise StopIteration
-        self.queryParams['cursorMark'] = self.nextCursorMark
-        self.get_next_results()
-        if self.nextCursorMark == self.queryParams['cursorMark']:
-            if self.index >= self.numFound:
-                raise StopIteration
-        if len(self.results['response']['docs']) == 0:
-            raise StopIteration
-        self.index += 1
-        return self.iter.next()
 
 
 class MARCFetcher(Fetcher):
